+++ Warcraft II/includes/RW/Settings.h
@@ -28,8 +28,8 @@ namespace RW
 		static const sf::Vector2i getRes() { return sf::Vector2i(jsonfile["resolution"]["width"].get<int>(), jsonfile["resolution"]["height"].get<int>()); }
 		static const sf::Vector2i getOldRes() { return oldres; }
 		static const sf::Vector2f get1920Scale() { return sf::Vector2f(jsonfile["resolution"]["width"].get<float>() / res1920.x, jsonfile["resolution"]["height"].get<float>() / res1920.y); }
-		static const unsigned short getMusic() { return jsonfile["music"].get<unsigned short>(); }
-		static const unsigned short getSoundFX() { return jsonfile["soundfx"].get<unsigned short>(); }
+		static const unsigned int getMusic() { return jsonfile["music"].get<unsigned int>(); }
+		static const unsigned int getSoundFX() { return jsonfile["soundfx"].get<unsigned int>(); }
 		static const std::string getName() { return jsonfile["playerName"].get<std::string>(); }
 		//holds 1/12th of the screens values
 		static const sf::Vector2f unit() { return sf::Vector2f((res1920.x/12) * get1920Scale().x, (res1920.y/12) * get1920Scale().y); }
@@ -39,8 +39,8 @@ namespace RW
 		static void setOldRes(const sf::Vector2i &oldres_in) { oldres = oldres_in; }
 		static void setFullscreen(const bool &fullscreen_in) { jsonfile["fullscreen"] = fullscreen_in; }
 		static void setOldFullscreen(const bool &oldfullscreen_in) { oldfullscreen = oldfullscreen_in; }
-		static void setMusic(const unsigned short &music_in) { jsonfile["music"] = music_in; }
-		static void setSoundFX(const unsigned short &soundfx_in) { jsonfile["soundfx"] = soundfx_in; }
+		static void setMusic(const unsigned int &music_in) { jsonfile["music"] = music_in; }
+		static void setSoundFX(const unsigned int &soundfx_in) { jsonfile["soundfx"] = soundfx_in; }
 		static void setName(const std::string & name_in) { jsonfile["playerName"] = name_in; }
 	};
 }
