+++ Warcraft II/includes/UI/Cursor.h
@@ -6,7 +6,7 @@ namespace UI
 	class Cursor : public sf::Drawable, public sf::Transformable
 	{
 	private:
-		const unsigned short size = 64;
+		const unsigned int size = 64;
 		sf::Texture cursorT;
 
 		sf::VertexArray vertices;
@@ -18,9 +18,9 @@ namespace UI
 
 			// define its 4 texture coordinates
 			vertices[0].texCoords = sf::Vector2f(0, 0);
-			vertices[1].texCoords = sf::Vector2f(size, 0);
-			vertices[2].texCoords = sf::Vector2f(size, size);
-			vertices[3].texCoords = sf::Vector2f(0, size);
+			vertices[1].texCoords = sf::Vector2f((float)size, 0);
+			vertices[2].texCoords = sf::Vector2f((float)size, (float)size);
+			vertices[3].texCoords = sf::Vector2f(0, (float)size);
 		}
 		void draw(sf::RenderTarget& target, sf::RenderStates states) const;
